requirements.txt
flask
requests
pandas
numpy
scikit-learn
sentence-transformers
faiss-cpu
torch
